requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sentence-transformers==3.0.1
tqdm==4.66.5
requests==2.32.3
torch==2.2.0